Move openapi v2 tests to its own package
f
