+++ http/codegen/openapi.go
@@ -1,12 +1,6 @@
 package codegen
 
 import (
-	"encoding/json"
-	"path/filepath"
-	"text/template"
-
-	"gopkg.in/yaml.v2"
-
 	"goa.design/goa/v3/codegen"
 	"goa.design/goa/v3/expr"
 	openapiv2 "goa.design/goa/v3/http/codegen/openapi/v2"
@@ -23,74 +17,18 @@ func OpenAPIFiles(root *expr.RootExpr) ([]*codegen.File, error) {
 	var files []*codegen.File
 	{
 		// OpenAPI v2
-		{
-			spec, err := openapiv2.NewV2(root, root.API.Servers[0].Hosts[0])
-			if err != nil {
-				return nil, err
-			}
-			jsonSection := &codegen.SectionTemplate{
-				Name:    "openapi",
-				FuncMap: template.FuncMap{"toJSON": toJSON},
-				Source:  "{{ toJSON .}}",
-				Data:    spec,
-			}
-			yamlSection := &codegen.SectionTemplate{
-				Name:    "openapi",
-				FuncMap: template.FuncMap{"toYAML": toYAML},
-				Source:  "{{ toYAML .}}",
-				Data:    spec,
-			}
-			files = []*codegen.File{
-				{
-					Path:             filepath.Join(codegen.Gendir, "http", "openapi.json"),
-					SectionTemplates: []*codegen.SectionTemplate{jsonSection},
-				},
-				{
-					Path:             filepath.Join(codegen.Gendir, "http", "openapi.yaml"),
-					SectionTemplates: []*codegen.SectionTemplate{yamlSection},
-				},
-			}
+		fs, err := openapiv2.Files(root)
+		if err != nil {
+			return nil, err
 		}
+		files = append(files, fs...)
+
 		// OpenAPI v3
-		{
-			spec := openapiv3.New(root)
-			jsonSection := &codegen.SectionTemplate{
-				Name:    "openapi_v3",
-				FuncMap: template.FuncMap{"toJSON": toJSON},
-				Source:  "{{ toJSON .}}",
-				Data:    spec,
-			}
-			yamlSection := &codegen.SectionTemplate{
-				Name:    "openapi_v3",
-				FuncMap: template.FuncMap{"toYAML": toYAML},
-				Source:  "{{ toYAML .}}",
-				Data:    spec,
-			}
-			files = append(files, &codegen.File{
-				Path:             filepath.Join(codegen.Gendir, "http", "openapi_v3.json"),
-				SectionTemplates: []*codegen.SectionTemplate{jsonSection},
-			})
-			files = append(files, &codegen.File{
-				Path:             filepath.Join(codegen.Gendir, "http", "openapi_v3.yaml"),
-				SectionTemplates: []*codegen.SectionTemplate{yamlSection},
-			})
+		fs, err = openapiv3.Files(root)
+		if err != nil {
+			return nil, err
 		}
+		files = append(files, fs...)
 	}
 	return files, nil
-}
-
-func toJSON(d interface{}) string {
-	b, err := json.Marshal(d)
-	if err != nil {
-		panic("openapi: " + err.Error()) // bug
-	}
-	return string(b)
-}
-
-func toYAML(d interface{}) string {
-	b, err := yaml.Marshal(d)
-	if err != nil {
-		panic("openapi: " + err.Error()) // bug
-	}
-	return string(b)
 }
