+++ http/codegen/openapi/v2/testdata/TestExtensions/endpoint_file0.golden
