+++ http/codegen/openapi/v2/testdata/TestExtensions/endpoint_file1.golden
