+++ http/codegen/openapi/v2/testdata/TestSections/empty_file1.golden
@@ -1,5 +1,6 @@
 swagger: "2.0"
 info:
+  title: ""
   version: ""
 host: goa.design
 consumes:
