+++ http/codegen/openapi/v2/testdata/TestSections/explicit-view_file0.golden
