+++ http/codegen/openapi/v2/testdata/TestSections/explicit-view_file1.golden
