+++ http/codegen/openapi/v2/testdata/TestSections/file-service_file0.golden
