+++ http/codegen/openapi/v2/testdata/TestSections/multiple-services_file0.golden
