+++ http/codegen/openapi/v2/testdata/TestSections/multiple-services_file1.golden
