+++ http/codegen/openapi/v2/testdata/TestSections/multiple-views_file0.golden
