+++ http/codegen/openapi/v2/testdata/TestSections/multiple-views_file1.golden
