+++ http/codegen/openapi/v2/testdata/TestSections/path-with-wildcards_file1.golden
