+++ http/codegen/openapi/v2/testdata/TestSections/security_file0.golden
