+++ http/codegen/openapi/v2/testdata/TestSections/server-host-with-variables_file0.golden
