+++ http/codegen/openapi/v2/testdata/TestSections/server-host-with-variables_file1.golden
