+++ http/codegen/openapi/v2/testdata/TestSections/valid_file1.golden
