+++ http/codegen/openapi/v2/testdata/TestSections/with-map_file0.golden
