+++ http/codegen/openapi/v2/testdata/TestSections/with-map_file1.golden
