+++ http/codegen/openapi/v2/testdata/TestSections/with-spaces_file0.golden
