+++ http/codegen/openapi/v2/testdata/TestSections/with-spaces_file1.golden
