+++ http/codegen/openapi/v2/testdata/TestValidations/array_file0.golden
