+++ http/codegen/openapi/v2/testdata/TestValidations/integer_file0.golden
