+++ http/codegen/openapi/v2/testdata/TestValidations/string_file0.golden
