+++ http/codegen/openapi/v2/testdata/TestValidations/string_file1.golden
