+++ http/codegen/testdata/openapi_dsls.go
